PMM-10059 CheckFilters
